+++ src/lsp/protocol.ts
@@ -361,6 +361,10 @@ export namespace CanShowMissingRequirementNotification {
   export const type = new lsp.RequestType<string, boolean, void>('sonarlint/canShowMissingRequirementsNotification');
 }
 
+export namespace DoNotShowMissingRequirementsMessageAgain {
+  export const type = new lsp.NotificationType('sonarlint/doNotShowMissingRequirementsMessageAgain');
+}
+
 export namespace MaybeShowWiderLanguageSupportNotification {
   export const type = new lsp.NotificationType<string[]>('sonarlint/maybeShowWiderLanguageSupportNotification');
 }
@@ -531,6 +535,15 @@ export namespace HelpAndFeedbackLinkClicked {
   );
 }
 
+export interface LMToolCalledNotificationParams {
+  toolName: string;
+  success: boolean;
+}
+
+export namespace LMToolCalled {
+  export const type = new lsp.NotificationType<LMToolCalledNotificationParams>('sonarlint/lmToolCalled');
+}
+
 export interface ScanFolderForHotspotsParams {
   folderUri: string;
   documents: Array<lsp.TextDocumentItem>;
@@ -572,6 +585,20 @@ export interface GetSuggestedBindingResponse {
 export namespace GetSuggestedBinding {
   export const type = new lsp.RequestType<GetSuggestedBindingParams, GetSuggestedBindingResponse, null>(
     'sonarlint/getBindingSuggestion'
+  );
+}
+
+export interface GetSuggestedConnectionsParams {
+  configurationScopeId: string;
+}
+
+export interface GetConnectionSuggestionsResponse {
+  connectionSuggestions: ConnectionSuggestion[];
+}
+
+export namespace GetSuggestedConnections {
+  export const type = new lsp.RequestType<GetSuggestedConnectionsParams, GetConnectionSuggestionsResponse, null>(
+    'sonarlint/getConnectionSuggestions'
   );
 }